src/requirements.txt
cloudshell-tg-breaking-point>=1.0,<1.1
cloudshell-networking-devices>=2.0,<2.1
cloudshell-shell-core>=3.1,<3.2
